requirements.txt
flask>=3.1.0
flask-cors>=5.0.0
google-genai>=1.0.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
pytz>=2024.1
